+++ packages/backend-data/src/factory.ts
@@ -17,11 +17,7 @@ import { GraphqlOutput } from '@aws-amplify/backend-output-schemas';
 import * as path from 'path';
 import { AmplifyDataError, DataProps } from './types.js';
 import { convertSchemaToCDK, isModelSchema } from './convert_schema.js';
-import {
-  FunctionInstanceProvider,
-  buildConstructFactoryFunctionInstanceProvider,
-  convertFunctionNameMapToCDK,
-} from './convert_functions.js';
+import { convertFunctionNameMapToCDK } from './convert_functions.js';
 import {
   ProvidedAuthConfig,
   buildConstructFactoryProvidedAuthConfig,
@@ -32,6 +28,11 @@ import { validateAuthorizationModes } from './validate_authorization_modes.js';
 import { AmplifyUserError, CDKContextKey } from '@aws-amplify/platform-core';
 import { Aspects, IAspect } from 'aws-cdk-lib';
 import { convertJsResolverDefinition } from './convert_js_resolvers.js';
+import { AppSyncPolicyGenerator } from './app_sync_policy_generator.js';
+import {
+  FunctionSchemaAccess,
+  JsResolver,
+} from '@aws-amplify/data-schema-types';
 
 /**
  * Singleton factory for AmplifyGraphqlApi constructs that can be used in Amplify project files.
@@ -82,7 +83,7 @@ export class DataFactory implements ConstructFactory<AmplifyData> {
             )
             ?.getInstance(props)
         ),
-        buildConstructFactoryFunctionInstanceProvider(props),
+        props,
         outputStorageStrategy
       );
     }
@@ -97,18 +98,52 @@ class DataGenerator implements ConstructContainerEntryGenerator {
   constructor(
     private readonly props: DataProps,
     private readonly providedAuthConfig: ProvidedAuthConfig | undefined,
-    private readonly functionInstanceProvider: FunctionInstanceProvider,
+    private readonly getInstanceProps: ConstructFactoryGetInstanceProps,
     private readonly outputStorageStrategy: BackendOutputStorageStrategy<GraphqlOutput>
   ) {}
 
-  generateContainerEntry = ({ scope }: GenerateContainerEntryProps) => {
+  generateContainerEntry = ({
+    scope,
+    ssmEnvironmentEntriesGenerator,
+  }: GenerateContainerEntryProps) => {
+    let amplifyGraphqlDefinition;
+    let jsFunctions: JsResolver[] = [];
+    let functionSchemaAccess: FunctionSchemaAccess[] = [];
+    try {
+      if (isModelSchema(this.props.schema)) {
+        ({ jsFunctions, functionSchemaAccess } = this.props.schema.transform());
+      }
+      amplifyGraphqlDefinition = convertSchemaToCDK(this.props.schema);
+    } catch (error) {
+      throw new AmplifyUserError<AmplifyDataError>(
+        'InvalidSchemaError',
+        {
+          message:
+            error instanceof Error
+              ? error.message
+              : 'Cannot covert user schema',
+        },
+        error instanceof Error ? error : undefined
+      );
+    }
+
     let authorizationModes;
+
+    /**
+     * TODO - remove this after the data construct does work to remove the need for allow-listed IAM roles
+     */
+    const functionSchemaAccessRoles = functionSchemaAccess.map(
+      (accessEntry) =>
+        accessEntry.resourceProvider.getInstance(this.getInstanceProps)
+          .resources.lambda.role!
+    );
 
     try {
       authorizationModes = convertAuthorizationModesToCDK(
-        this.functionInstanceProvider,
+        this.getInstanceProps,
         this.providedAuthConfig,
-        this.props.authorizationModes
+        this.props.authorizationModes,
+        functionSchemaAccessRoles
       );
     } catch (error) {
       throw new AmplifyUserError<AmplifyDataError>(
@@ -147,30 +182,9 @@ class DataGenerator implements ConstructContainerEntryGenerator {
     );
 
     const functionNameMap = convertFunctionNameMapToCDK(
-      this.functionInstanceProvider,
+      this.getInstanceProps,
       this.props.functions ?? {}
     );
-
-    let amplifyGraphqlDefinition;
-    let jsFunctions;
-    try {
-      if (isModelSchema(this.props.schema)) {
-        ({ jsFunctions } = this.props.schema.transform());
-      }
-      amplifyGraphqlDefinition = convertSchemaToCDK(this.props.schema);
-    } catch (error) {
-      throw new AmplifyUserError<AmplifyDataError>(
-        'InvalidSchemaError',
-        {
-          message:
-            error instanceof Error
-              ? error.message
-              : 'Cannot covert user schema',
-        },
-        error instanceof Error ? error : undefined
-      );
-    }
-
     const amplifyApi = new AmplifyData(scope, this.defaultName, {
       apiName: this.props.name,
       definition: amplifyGraphqlDefinition,
@@ -198,6 +212,26 @@ class DataGenerator implements ConstructContainerEntryGenerator {
     }
 
     convertJsResolverDefinition(scope, amplifyApi, jsFunctions);
+
+    const ssmEnvironmentEntries =
+      ssmEnvironmentEntriesGenerator.generateSsmEnvironmentEntries({
+        [`${this.props.name}_GRAPHQL_ENDPOINT`]:
+          amplifyApi.resources.cfnResources.cfnGraphqlApi.attrGraphQlUrl,
+      });
+
+    const policyGenerator = new AppSyncPolicyGenerator(
+      amplifyApi.resources.graphqlApi
+    );
+
+    functionSchemaAccess.forEach((accessDefinition) => {
+      const policy = policyGenerator.generateGraphqlAccessPolicy(
+        accessDefinition.actions
+      );
+      accessDefinition.resourceProvider
+        .getInstance(this.getInstanceProps)
+        .getResourceAccessAcceptor()
+        .acceptResourceAccess(policy, ssmEnvironmentEntries);
+    });
 
     return amplifyApi;
   };