+++ packages/integration-tests/src/test-project-setup/data_storage_auth_with_triggers.ts
@@ -103,6 +103,8 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
 
   private readonly dataResourceFileSuffix = 'data/resource.ts';
 
+  private readonly functionHandlerFileSuffix = 'func-src/handler.ts';
+
   private readonly testSecretNames = [
     'googleId',
     'googleSecret',
@@ -174,12 +176,30 @@ class DataStorageAuthWithTriggerTestProject extends TestProjectBase {
     const dataResourceFile = pathToFileURL(
       path.join(this.projectAmplifyDirPath, this.dataResourceFileSuffix)
     );
+
+    const sourceFunctionUpdateFile = pathToFileURL(
+      path.join(
+        fileURLToPath(this.sourceProjectUpdateDirPath),
+        this.functionHandlerFileSuffix
+      )
+    );
+    const functionHandlerFile = pathToFileURL(
+      path.join(this.projectAmplifyDirPath, this.functionHandlerFileSuffix)
+    );
     return [
       {
         sourceFile: sourceDataResourceFile,
         projectFile: dataResourceFile,
         deployThresholdSec: {
           onWindows: 40,
+          onOther: 30,
+        },
+      },
+      {
+        sourceFile: sourceFunctionUpdateFile,
+        projectFile: functionHandlerFile,
+        deployThresholdSec: {
+          onWindows: 30,
           onOther: 30,
         },
       },