+++ packages/auth-construct/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @aws-amplify/auth-construct-alpha
 
+## 0.6.0-beta.4
+
+### Patch Changes
+
+- 1d444df: Fix deployment bug with SAML providers.
+
 ## 0.6.0-beta.3
 
 ### Patch Changes